Create a p-value functions object that can be used to generate p-values

Calculating p-values is not trivial for ALE statistics because ALE is non-parametric and model-agnostic. Because ALE is non-parametric (that is, it does not assume any particular distribution of data), the ale package generates p-values by calculating ALE for many random variables; this makes the procedure somewhat slow. For this reason, they are not calculated by default; they must be explicitly requested. Because the ale package is model-agnostic (that is, it works with any kind of R model), the ale() function cannot always automatically manipulate the model object to create the p-values. It can only do so for models that follow the standard R statistical modelling conventions, which includes almost all built-in R algorithms (like stats::lm() and stats::glm()) and many widely used statistics packages (like mgcv and survival), but which excludes most machine learning algorithms (like tidymodels and caret). For non-standard algorithms, the user needs to do a little work to help the ale function correctly manipulate its model object.

Usage

create_p_funs(
  training_data,
  test_data,
  model,
  ...,
  parallel = parallel::detectCores(logical = FALSE) - 1,
  model_packages = as.character(NA),
  random_model_call_string = NULL,
  random_model_call_string_vars = character(),
  y_col = NULL,
  pred_fun = function(object, newdata, type = pred_type) {
     stats::predict(object =
    object, newdata = newdata, type = type)
 },
  pred_type = "response",
  rand_it = 1000,
  silent = FALSE,
  .testing_mode = FALSE
)

Arguments

training_data

dataframe. The dataset originally used to train model. See details and also documentation for ale().

test_data

dataframe. The dataset that will be used to create ALE data. See details.

model

model object. The model used to train the original training_data. for which ALE should be calculated. See details and also documentation for ale().

...

not used. Inserted to require explicit naming of subsequent arguments.

parallel

See documentation for ale()

model_packages

See documentation for ale()

random_model_call_string

character string. If NULL, create_p_funs() tries to automatically detect and construct the call for p-values. If it cannot, the function will fail early. In that case, a character string of the full call for the model must be provided that includes the random variable. See details.

random_model_call_string_vars

See documentation for model_call_string_vars in model_bootstrap(); the operation is very similar.

y_col

See documentation for ale()

pred_fun, pred_type

See documentation for ale().

rand_it

non-negative integer length 1. Number of times that the model should be retrained with a new random variable. The default of 1000 should give reasonably stable p-values. It can be reduced as low as 100 for faster test runs.

silent

See documentation for ale()

.testing_mode

logical. Internal use only.

Value

The return value is a list of class c('p_funs', 'ale', 'list') with an ale_version attribute whose value is the version of the ale package used to create the object. See examples for an illustration of how to inspect this list. Its elements are:

value_to_p: a list of functions named for each each available ALE statistic. Each function signature is function(x) where x is a numeric. The function returns the p-value (minimum 0; maximum 1) for the respective statistic based on the random variable analysis. For an input x that returns p, its interpretation is that p% of random variables obtained the same or higher statistic value. For example, to get the p-value of a NALED of 4.2, enter p_funs$value_to_p(4.2). A return value of 0.03 means that only 3% of random variables obtained a NALED greater than or equal to 4.2.

p_to_random_value: a list of functions named for each each available ALE statistic. These are the inverse functions of value_to_p. The signature is function(p) where p is a numeric from 0 to 1. The function returns the numeric value of the random variable statistic that would yield the provided p-value. For an input p that returns x, its interpretation is that p% of random variables obtained the same or higher statistic value. For example, to get the random variable ALED for the 0.05 p-value, enter p_funs$p_to_random_value(0.05). A return value of 102 means that only 5% of random variables obtained an ALED greater than or equal to 102.

rand_stats: a tibble whose rows are each of the rand_it iterations of the random variable analysis and whose columns are the ALE statistics obtained for each random variable.

residuals: the actual y_col values from training_data minus the predicted values from the model (without random variables) on the training_data. residual_distribution: the closest estimated distribution for the residuals as determined by univariateML::rml(). This is the distribution used to generate all the random variables.

Approach to calculating p-values

The ale package takes a literal frequentist approach to the calculation of p-values. That is, it literally retrains the model 1000 times, each time modifying it by adding a distinct random variable to the model. (The number of iterations is customizable with the rand_it argument.) The ALEs and ALE statistics are calculated for each random variable. The percentiles of the distribution of these random-variable ALEs are then used to determine p-values for non-random variables. Thus, p-values are interpreted as the frequency of random variable ALE statistics that exceed the value of ALE statistic of the actual variable in question. The specific steps are as follows:

The residuals of the original model trained on the training data are calculated (residuals are the actual y target value minus the predicted values).

The closest distribution of the residuals is detected with univariateML::model_select().

1000 new models are trained by generating a random variable each time with univariateML::rml() and then training a new model with that random variable added.

The ALEs and ALE statistics are calculated for each random variable.

For each ALE statistic, the empirical cumulative distribution function (from stats::ecdf()) is used to create a function to determine p-values according to the distribution of the random variables' ALE statistics.

Datasets

Because the ale package takes a literal frequentist approach to the calculation of p-values, the precision of the p-values depends on correctly representing the modelling workflow. Specifically, models should be trained on a training subset and ALE statistics should be calculated on a test subset. Thus, the random variables should be trained on the same training subset and their ALE statistics should be calculated on the same test subset. Thus the model takes both training_data and test_data as distinct arguments. If, for whatever reason, the ALE is calculated on the same dataset as is used to train the model, then the same dataset should be entered for both training_data and test_data.

References

Okoli, Chitu. 2023. “Statistical Inference Using Machine Learning and Classical Techniques Based on Accumulated Local Effects (ALE).” arXiv. https://arxiv.org/abs/2310.09877.

Examples

# \donttest{
# Sample 1000 rows from the diamonds dataset (for a simple example)diamonds
set.seed(0)
diamonds_sample <- diamonds[sample(nrow(diamonds), 1000), ]

# Split the dataset into training and test sets
# https://stackoverflow.com/a/54892459/2449926
set.seed(0)
train_test_split <- sample(
  c(TRUE, FALSE), nrow(diamonds_sample), replace = TRUE, prob = c(0.8, 0.2)
)
diamonds_train <- diamonds_sample[train_test_split, ]
diamonds_test <- diamonds_sample[!train_test_split, ]

# Create a GAM model with flexible curves to predict diamond price
# Smooth all numeric variables and include all other variables
# Build model on training data, not on the full dataset.
gam_diamonds <- mgcv::gam(
  price ~ s(carat) + s(depth) + s(table) + s(x) + s(y) + s(z) +
    cut + color + clarity,
  data = diamonds_train
)
summary(gam_diamonds)
#> 
#> Family: gaussian 
#> Link function: identity 
#> 
#> Formula:
#> price ~ s(carat) + s(depth) + s(table) + s(x) + s(y) + s(z) + 
#>     cut + color + clarity
#> 
#> Parametric coefficients:
#>               Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  3311.6596    82.3642  40.208  < 2e-16 ***
#> cut.L         226.8516   174.6866   1.299  0.19447    
#> cut.Q         154.4894   136.7646   1.130  0.25900    
#> cut.C         -72.0299   114.5065  -0.629  0.52951    
#> cut^4          63.4426    91.1593   0.696  0.48667    
#> color.L     -1748.6107   119.8822 -14.586  < 2e-16 ***
#> color.Q      -499.5061   109.9114  -4.545 6.41e-06 ***
#> color.C        -7.7188   102.4915  -0.075  0.93999    
#> color^4       132.4433    95.3721   1.389  0.16533    
#> color^5      -261.4225    89.7563  -2.913  0.00369 ** 
#> color^6        34.2734    81.9067   0.418  0.67574    
#> clarity.L    4464.0110   271.7651  16.426  < 2e-16 ***
#> clarity.Q   -3168.1893   261.5197 -12.115  < 2e-16 ***
#> clarity.C    1276.9741   215.1091   5.936 4.44e-09 ***
#> clarity^4    -983.6948   161.9254  -6.075 1.97e-09 ***
#> clarity^5     557.2408   120.8123   4.612 4.67e-06 ***
#> clarity^6    -117.3182   100.4726  -1.168  0.24331    
#> clarity^7      -0.9172    87.9078  -0.010  0.99168    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Approximate significance of smooth terms:
#>            edf Ref.df     F  p-value    
#> s(carat) 8.507  8.820 3.423 0.000389 ***
#> s(depth) 1.273  1.510 0.061 0.901571    
#> s(table) 5.707  6.842 1.522 0.207715    
#> s(x)     5.006  6.125 4.761 8.88e-05 ***
#> s(y)     5.644  6.755 9.317  < 2e-16 ***
#> s(z)     2.516  3.307 1.128 0.335376    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> R-sq.(adj) =  0.942   Deviance explained = 94.5%
#> GCV = 9.5069e+05  Scale est. = 8.9531e+05  n = 801

# Create p-value functions
pf_diamonds <- create_p_funs(
  diamonds_train,
  diamonds_test,
  gam_diamonds,
  # only 100 iterations for a quick demo; but usually should remain at 1000
  rand_it = 100,
  model_packages = 'mgcv',  # required for parallel processing
)

# Examine the structure of the returned object
str(pf_diamonds)
#> List of 5
#>  $ value_to_p           :List of 6
#>   ..$ aled     :function (x)  
#>   ..$ aler_min :function (x)  
#>   ..$ aler_max :function (x)  
#>   ..$ naled    :function (x)  
#>   ..$ naler_min:function (x)  
#>   ..$ naler_max:function (x)  
#>  $ p_to_random_value    :List of 6
#>   ..$ aled     :function (p)  
#>   ..$ aler_min :function (p)  
#>   ..$ aler_max :function (p)  
#>   ..$ naled    :function (p)  
#>   ..$ naler_min:function (p)  
#>   ..$ naler_max:function (p)  
#>  $ rand_stats           : tibble [100 × 6] (S3: tbl_df/tbl/data.frame)
#>   ..$ aled     : num [1:100] 25.58 11.85 6.25 30.87 16.93 ...
#>   ..$ aler_min : num [1:100] -180.2 -57.9 -17.1 -124.9 -98 ...
#>   ..$ aler_max : num [1:100] 122 42.5 28.5 214.6 171.4 ...
#>   ..$ naled    : num [1:100] 0.3208 0.1385 0.0401 0.3853 0.2162 ...
#>   ..$ naler_min: num [1:100] -2.868 -1.122 -0.374 -1.995 -1.87 ...
#>   ..$ naler_max: num [1:100] 0.873 0.249 0.249 1.621 1.247 ...
#>  $ residuals            : num [1:801(1d)] -1180.2 -1096.3 -216.6 -25.6 -1511.5 ...
#>  $ residual_distribution: 'univariateML' Named num [1:4] 9.71 1020.87 2.99 1.24
#>   ..- attr(*, "names")= chr [1:4] "mean" "sd" "nu" "xi"
#>   ..- attr(*, "model")= chr "Skew Student-t"
#>   ..- attr(*, "density")= chr "fGarch::dsstd"
#>   ..- attr(*, "logLik")= num -6503
#>   ..- attr(*, "support")= num [1:2] -Inf Inf
#>   ..- attr(*, "n")= int 801
#>   ..- attr(*, "call")= language f(x = x, na.rm = na.rm)
#>  - attr(*, "class")= chr [1:3] "p_funs" "ale" "list"
#>  - attr(*, "ale_version")=Classes 'package_version', 'numeric_version'  hidden list of 1
#>   ..$ : int [1:4] 0 2 0 20240206
# In RStudio: View(pf_diamonds)

# Calculate ALEs with p-values
ale_gam_diamonds <- ale(
  diamonds_test,
  gam_diamonds,
  model_packages = 'mgcv',  # required for parallel processing
  p_values = pf_diamonds
)
#> Error in ale(diamonds_test, gam_diamonds, model_packages = "mgcv", p_values = pf_diamonds): object 'diamonds_test' not found

# Plot the ALE data. The horizontal bands in the plots use the p-values.
gridExtra::grid.arrange(grobs = ale_gam_diamonds$plots, ncol = 2)
#> Error in eval(expr, envir, enclos): object 'ale_gam_diamonds' not found

# }