Improvements:

Validation: ensure that the object is atomic (not just a vector)

For factors, get all classes and check if any class_x is a factor or ordered

Add arguments to return a unique mode with options to sort: occurrence order, lexicographical Reduce a dataframe to a sample (retains the structure of its columns)

Improvements:

Validation: ensure that the object is atomic (not just a vector)

For factors, get all classes and check if any class_x is a factor or ordered

Add arguments to return a unique mode with options to sort: occurrence order, lexicographical Reduce a dataframe to a sample (retains the structure of its columns)

Usage

params_data(
  data,
  y_vals,
  data_name = var_name(data),
  sample_size = 500,
  seed = 0
)

Arguments

data

input dataframe

y_vals

y values, y predictions, or a sample thereof

data_name

name of the data argument

sample_size

size of data to sample

seed

random seed

Value

a list