Create and return ALE data, statistics, and plots

ale() is the central function that manages the creation of ALE data and plots for one-way ALE. For two-way interactions, see ale_ixn(). This function calls ale_core (a non-exported function) that manages the ALE data and plot creation in detail. For details, see the introductory vignette for this package or the details and examples below.

Custom predict function

The calculation of ALE requires modifying several values of the original data. Thus, ale() needs direct access to a predict function that work on model. By default, ale() uses a generic default predict function of the form predict(object, newdata, type) with the default prediction type of 'response'. If, however, the desired prediction values are not generated with that format, the user must specify what they want. Most of the time, the only modification needed is to change the prediction type to some other value by setting the pred_type argument (e.g., to 'prob' to generated classification probabilities). But if the desired predictions need a different function signature, then the user must create a custom prediction function and pass it to pred_fun. The requirements for this custom function are:

It must take three required arguments and nothing else:

object: a model

newdata: a dataframe or compatible table type

type: a string; it should usually be specified as type = pred_type These argument names are according to the R convention for the generic stats::predict function.

It must return a vector of numeric values as the prediction.

You can see an example below of a custom prediction function.

Note: survival models probably do not need a custom prediction function but y_col must be set to the name of the binary event column and pred_type must be set to the desired prediction type.

ALE statistics

For details about the ALE-based statistics (ALED, ALER, NALED, and NALER), see vignette("ale-statistics").

About the ale package

Accumulated Local Effects (ALE) were initially developed as a model-agnostic approach for global explanations of the results of black-box machine learning algorithms. ALE has a key advantage over other approaches like partial dependency plots (PDP) and SHapley Additive exPlanations (SHAP): its values represent a clean functional decomposition of the model. As such, ALE values are not affected by the presence or absence of interactions among variables in a mode. Moreover, its computation is relatively rapid. This package rewrites the original code from the 'ALEPlot' package for calculating ALE data and it completely reimplements the plotting of ALE values. It also extends the original ALE concept to add bootstrap-based confidence intervals and ALE-based statistics that can be used for statistical inference. For more details, see Okoli, Chitu. 2023. “Statistical Inference Using Machine Learning and Classical Techniques Based on Accumulated Local Effects (ALE).” arXiv. https://arxiv.org/abs/2310.09877.

Usage

ale(
  data,
  model,
  x_cols = NULL,
  y_col = NULL,
  ...,
  output = c("plots", "data", "stats"),
  pred_fun = function(object, newdata, type = pred_type) {
     stats::predict(object =
    object, newdata = newdata, type = type)
 },
  pred_type = "response",
  p_values = NULL,
  x_intervals = 100,
  boot_it = 0,
  seed = 0,
  boot_alpha = 0.05,
  boot_centre = "mean",
  relative_y = "median",
  y_type = NULL,
  median_band = c(0.05, 0.5),
  rug_sample_size = 500,
  min_rug_per_interval = 1,
  ale_xs = NULL,
  ale_ns = NULL,
  silent = FALSE
)

Arguments

data

dataframe. Dataset from which to create predictions for the ALE.

model

model object. Model for which ALE should be calculated. May be any kind of R object that can make predictions from data.

x_cols

character. Vector of column names from data for which one-way ALE data is to be calculated (that is, simple ALE without interactions). If not provided, ALE will be created for all columns in data except y_col.

y_col

character length 1. Name of the outcome target label (y) variable. If not provided, ale() will try to detect it automatically. For non-standard models, y_col should be provided. For survival models, set y_col to the name of the binary event column; in that case, pred_type should also be specified.

...

not used. Inserted to require explicit naming of subsequent arguments.

output

character in c('plots', 'data', 'stats'). Vector of types of results to return. 'plots' will return an ALE plot; 'data' will return the source ALE data; 'stats' will return ALE statistics. Each option must be listed to return the specified component. By default, all are returned.

pred_fun, pred_type

function,character length 1. pred_fun is a function that returns a vector of predicted values of type pred_type from model on data. See details.

p_values

instructions for calculating p-values. If NULL (default), no p-values are calculated. To calculate p-values, an object generated by the create_p_funs() function must be provided here. If p_values is set to 'auto', this ale() function will try to automatically create the p-values function; this only works with standard R model types. Any error message will be given if p-values cannot be generated. Any other input provided to this argument will result in an error. For more details about creating p-values, see documentation for create_p_funs(). Note that p-values will not be generated if 'stats' are not included as an option in the output argument.

x_intervals

positive integer length 1. Maximum number of intervals on the x-axis for the ALE data for each column in x_cols. The number of intervals that the algorithm generates might eventually be fewer than what the user specifies if the data values for a given x value do not support that many intervals.

boot_it

non-negative integer length 1. Number of bootstrap iterations for the ALE values. If boot_it = 0 (default), then ALE will be calculated on the entire dataset with no bootstrapping.

seed

integer length 1. Random seed. Supply this between runs to assure that identical random ALE data is generated each time

boot_alpha

numeric length 1 from 0 to 1. Alpha for percentile-based confidence interval range for the bootstrap intervals; the bootstrap confidence intervals will be the lowest and highest (1 - 0.05) / 2 percentiles. For example, if boot_alpha = 0.05 (default), the intervals will be from the 2.5 and 97.5 percentiles.

boot_centre

character length 1 in c('mean', 'median'). When bootstrapping, the main estimate for ale_y is considered to be boot_centre. Regardless of the value specified here, both the mean and median will be available.

relative_y

character length 1 in c('median', 'mean', 'zero'). The ale_y values will be adjusted relative to this value. 'median' is the default. 'zero' will maintain the default of ALEPlot::ALEPlot, which is not shifted.

y_type

character length 1. Datatype of the y (outcome) variable. Must be one of c('binary', 'numeric', 'multinomial', 'ordinal'). Normally determined automatically; only provide for complex non-standard models that require it.

median_band

numeric length 2 from 0 to 1. Alpha for "confidence interval" ranges for printing bands around the median for single-variable plots. The inner band range will be the median value of y ± median_band[1]. For plots with a second outer band, its range will be the median ± median_band[2].

rug_sample_size, min_rug_per_interval

single non-negative integer length 1. Rug plots are normally down-sampled otherwise they are too slow. rug_sample_size specifies the size of this sample. To prevent down-sampling, set to Inf. To suppress rug plots, set to 0. When down-sampling, the rug plots maintain representativeness of the data by guaranteeing that each of the x_intervals intervals will retain at least min_rug_per_interval elements; usually set to just 1 or 2.

ale_xs, ale_ns

list of ale_x and ale_n vectors. If provided, these vectors will be used to set the intervals of the ALE x axis for each variable. By default (NULL), the function automatically calculates the ale_x intervals. ale_xs is normally used in advanced analyses where the ale_x intervals from a previous analysis are reused for subsequent analyses (for example, for full model bootstrapping; see the model_bootstrap() function).

silent

logical length 1, default FALSE. If TRUE, do not display any non-essential messages during execution (such as progress bars). Regardless, any warnings and errors will always display.

Value

list with elements data, plots, and stats as requested in the output argument. Each of these is a list named by the x variables with the respective values for each variable. If any of these was not requested in the output argument, its return value will be NULL. In addition, the return object recapitulates several elements that were passed as arguments that apply to all the x variables for the ALE calculation.

Details

ale_core.R

Core functions for the ale package: ale, ale_ixn, and ale_core

References

Okoli, Chitu. 2023. “Statistical Inference Using Machine Learning and Classical Techniques Based on Accumulated Local Effects (ALE).” arXiv. https://arxiv.org/abs/2310.09877.

See also

Author

Chitu Okoli Chitu.Okoli@skema.edu

Examples

diamonds
#> # A tibble: 53,940 × 10
#>    carat cut       color clarity depth table price     x     y     z
#>    <dbl> <ord>     <ord> <ord>   <dbl> <dbl> <int> <dbl> <dbl> <dbl>
#>  1  0.23 Ideal     E     SI2      61.5    55   326  3.95  3.98  2.43
#>  2  0.21 Premium   E     SI1      59.8    61   326  3.89  3.84  2.31
#>  3  0.23 Good      E     VS1      56.9    65   327  4.05  4.07  2.31
#>  4  0.29 Premium   I     VS2      62.4    58   334  4.2   4.23  2.63
#>  5  0.31 Good      J     SI2      63.3    58   335  4.34  4.35  2.75
#>  6  0.24 Very Good J     VVS2     62.8    57   336  3.94  3.96  2.48
#>  7  0.24 Very Good I     VVS1     62.3    57   336  3.95  3.98  2.47
#>  8  0.26 Very Good H     SI1      61.9    55   337  4.07  4.11  2.53
#>  9  0.22 Fair      E     VS2      65.1    61   337  3.87  3.78  2.49
#> 10  0.23 Very Good H     VS1      59.4    61   338  4     4.05  2.39
#> # ℹ 53,930 more rows
set.seed(0)
diamonds_sample <- diamonds[sample(nrow(diamonds), 1000), ]

# Split the dataset into training and test sets
# https://stackoverflow.com/a/54892459/2449926
set.seed(0)
train_test_split <- sample(
  c(TRUE, FALSE), nrow(diamonds_sample), replace = TRUE, prob = c(0.8, 0.2)
)
diamonds_train <- diamonds_sample[train_test_split, ]
diamonds_test <- diamonds_sample[!train_test_split, ]

# Create a GAM model with flexible curves to predict diamond price
# Smooth all numeric variables and include all other variables
# Build model on training data, not on the full dataset.
gam_diamonds <- mgcv::gam(
  price ~ s(carat) + s(depth) + s(table) + s(x) + s(y) + s(z) +
    cut + color + clarity,
  data = diamonds_train
)
summary(gam_diamonds)
#> 
#> Family: gaussian 
#> Link function: identity 
#> 
#> Formula:
#> price ~ s(carat) + s(depth) + s(table) + s(x) + s(y) + s(z) + 
#>     cut + color + clarity
#> 
#> Parametric coefficients:
#>               Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  3311.6596    82.3642  40.208  < 2e-16 ***
#> cut.L         226.8516   174.6866   1.299  0.19447    
#> cut.Q         154.4894   136.7646   1.130  0.25900    
#> cut.C         -72.0299   114.5065  -0.629  0.52951    
#> cut^4          63.4426    91.1593   0.696  0.48667    
#> color.L     -1748.6107   119.8822 -14.586  < 2e-16 ***
#> color.Q      -499.5061   109.9114  -4.545 6.41e-06 ***
#> color.C        -7.7188   102.4915  -0.075  0.93999    
#> color^4       132.4433    95.3721   1.389  0.16533    
#> color^5      -261.4225    89.7563  -2.913  0.00369 ** 
#> color^6        34.2734    81.9067   0.418  0.67574    
#> clarity.L    4464.0110   271.7651  16.426  < 2e-16 ***
#> clarity.Q   -3168.1893   261.5197 -12.115  < 2e-16 ***
#> clarity.C    1276.9741   215.1091   5.936 4.44e-09 ***
#> clarity^4    -983.6948   161.9254  -6.075 1.97e-09 ***
#> clarity^5     557.2408   120.8123   4.612 4.67e-06 ***
#> clarity^6    -117.3182   100.4726  -1.168  0.24331    
#> clarity^7      -0.9172    87.9078  -0.010  0.99168    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Approximate significance of smooth terms:
#>            edf Ref.df     F  p-value    
#> s(carat) 8.507  8.820 3.423 0.000389 ***
#> s(depth) 1.273  1.510 0.061 0.901571    
#> s(table) 5.707  6.842 1.522 0.207715    
#> s(x)     5.006  6.125 4.761 8.88e-05 ***
#> s(y)     5.644  6.755 9.317  < 2e-16 ***
#> s(z)     2.516  3.307 1.128 0.335376    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> R-sq.(adj) =  0.942   Deviance explained = 94.5%
#> GCV = 9.5069e+05  Scale est. = 8.9531e+05  n = 801

# Simple ALE without bootstrapping
ale_gam_diamonds <- ale(diamonds_test, gam_diamonds)
#> Error in ale(diamonds_test, gam_diamonds): object 'diamonds_test' not found

# \donttest{
# Plot the ALE data
gridExtra::grid.arrange(grobs = ale_gam_diamonds$plots, ncol = 2)
#> Error in eval(expr, envir, enclos): object 'ale_gam_diamonds' not found

# Bootstrapped ALE
# This can be slow, since bootstrapping runs the algorithm boot_it times

# Create ALE with 100 bootstrap samples
ale_gam_diamonds_boot <- ale(diamonds_test, gam_diamonds, boot_it = 100)
#> Error in ale(diamonds_test, gam_diamonds, boot_it = 100): object 'diamonds_test' not found

# Bootstrapped ALEs print with confidence intervals
gridExtra::grid.arrange(grobs = ale_gam_diamonds_boot$plots, ncol = 2)
#> Error in eval(expr, envir, enclos): object 'ale_gam_diamonds_boot' not found

# If the predict function you want is non-standard, you may define a
# custom predict function. It must return a single numeric vector.
custom_predict <- function(object, newdata, type = pred_type) {
  predict(object, newdata, type = type, se.fit = TRUE)$fit
}

ale_gam_diamonds_custom <- ale(
  diamonds_test, gam_diamonds,
  pred_fun = custom_predict, pred_type = 'link'
)
#> Error in ale(diamonds_test, gam_diamonds, pred_fun = custom_predict, pred_type = "link"): object 'diamonds_test' not found

# Plot the ALE data
gridExtra::grid.arrange(grobs = ale_gam_diamonds_custom$plots, ncol = 2)
#> Error in eval(expr, envir, enclos): object 'ale_gam_diamonds_custom' not found

# }